backend/app/requirements.txt
alembic==1.4.2
APScheduler==3.6.3
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
bandit==1.6.2
bcrypt==3.1.7
blinker==1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
coverage==5.0.4
cryptography==2.9.2
dnspython==1.16.0
dodgy==0.2.1
email-validator==1.1.0
eventlet==0.25.2
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.1
Flask-User==1.0.2.2
Flask-WTF==0.14.3
gevent==20.5.2
gitdb==4.0.2
GitPython==3.1.0
greenlet==0.4.15
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jsonify==0.5
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
mollie-api-python==2.3.1
monotonic==1.5
more-itertools==8.2.0
mysql-connector-python==8.0.19
oauthlib==3.1.0
packaging==20.3
passlib==1.7.2
pbr==5.4.4
pep8-naming==0.4.1
pluggy==0.13.1
prospector==1.2.0
protobuf==3.6.1
psycopg2==2.8.5
py==1.8.1
pycodestyle==2.4.0
pycparser==2.20
pydocstyle==5.0.2
pyflakes==2.1.1
pylint==2.4.4
pylint-celery==0.3
pylint-django==2.0.12
pylint-flask==0.6
pylint-plugin-utils==0.6
pyparsing==2.4.6
pytest==5.4.1
pytest-cov==2.8.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-engineio==3.13.0
python-socketio==4.6.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
requirements-detector==0.6
setoptconf==0.2.0
six==1.14.0
smmap==3.0.1
snowballstemmer==2.0.0
SQLAlchemy==1.3.15
stevedore==1.32.0
stripe==2.48.0
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.11.2
WTForms==2.3.1
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
